fix(CPEX-0045): fix multiplication overflow and %ld format mismatch in lagrange alloc
Cast nnodes to size_t before multiplying to prevent int overflow before
the result is widened, and switch the error message from %ld/(long) to
%zu/(size_t) to match the actual type of the size expression.

+++ src/cgnslib.c
@@ -17495,10 +17495,10 @@ int cg_element_interpolation_points_write(int fn, int bn, int fam, int en ,
     einterp->lagrangePts->convert=0;
     einterp->lagrangePts->data=0;
     
-    einterp->lagrangePts->data = malloc( (size_t) ( nnodes * edim * sizeof(double) ) );
+    einterp->lagrangePts->data = malloc( (size_t)nnodes * edim * sizeof(double) );
     if (!einterp->lagrangePts->data) {
-        cgi_error("Error allocating %ld bytes for LagrangeControlPoints data",
-                  (long)(nnodes * edim * sizeof(double)));
+        cgi_error("Error allocating %zu bytes for LagrangeControlPoints data",
+                  (size_t)nnodes * edim * sizeof(double));
         CGNS_FREE(einterp->lagrangePts);
         einterp->lagrangePts = 0;
         return CG_ERROR;
@@ -18281,10 +18281,10 @@ int cg_solution_interpolation_points_write(int fn, int bn, int fam, int sn ,
     sinterp->lagrangePts->convert=0;
     sinterp->lagrangePts->data=0;
     
-    sinterp->lagrangePts->data = malloc( (size_t) ( nnodes * (edim + (ot ? 1 : 0)) * sizeof(double) ) );
+    sinterp->lagrangePts->data = malloc( (size_t)nnodes * (edim + (ot ? 1 : 0)) * sizeof(double) );
     if (!sinterp->lagrangePts->data) {
-        cgi_error("Error allocating %ld bytes for LagrangeControlPoints data",
-                  (long)(nnodes * (edim + (ot ? 1 : 0)) * sizeof(double)));
+        cgi_error("Error allocating %zu bytes for LagrangeControlPoints data",
+                  (size_t)nnodes * (edim + (ot ? 1 : 0)) * sizeof(double));
         CGNS_FREE(sinterp->lagrangePts);
         sinterp->lagrangePts = 0;
         return CG_ERROR;
